Implements drink choosing algorithm
Implements an algorithm to determine the maximum number of
drinks that can be consumed given a number of drinks and a
constraint on the types of drinks. The algorithm counts the
frequency of each drink type, and calculates the total number
of drinks that can be consumed based on the constraint.